Fix compiler warnings and vulnerable dependency
- Declare varint loop variable as long to fix CS0675 sign-extension warnings
- Remove dead break statements after return in GetIntegerValue
- Remove redundant no-op span slices in GetIntegerValue
- Override transitive Tmds.DBus.Protocol 0.20.0 with 0.94.1 to resolve NU1903

+++ SHARD.Core/Decoding/VarintDecoder.cs
@@ -19,7 +19,7 @@ public sealed class Varint
         long tempValue = 0;
         for (int i = 0; i < MAX_VARINT_LENGTH; i++)
         {
-            int current = paddedData[i];
+            long current = paddedData[i];
             //if greater than or equal to, more bytes to come
             if (current >= 128)
             {
